chore(structure): organize files into view, domain, and service packages
