+++ src/main/kotlin/lotto/domain/Rank.kt
@@ -1,0 +1,4 @@
+package lotto.domain
+
+class Rank {
+}
