+++ src/main/kotlin/lotto/domain/WinningNumbers.kt
@@ -1,4 +1,6 @@
-package lotto
+package lotto.domain
+
+import lotto.Lotto
 
 class WinningNumbers(
     private val numbers: List<Int>,
